delete hooks, merge merkle tree functionality into mailbox
