+++ dango/cli/src/start.rs
@@ -62,10 +62,8 @@ impl StartCmd {
             codes.account_spot.to_bytes().hash256(),
             codes.bank.to_bytes().hash256(),
             codes.dex.to_bytes().hash256(),
-            codes.hyperlane.fee.to_bytes().hash256(),
             codes.hyperlane.ism.to_bytes().hash256(),
             codes.hyperlane.mailbox.to_bytes().hash256(),
-            codes.hyperlane.merkle.to_bytes().hash256(),
             codes.hyperlane.va.to_bytes().hash256(),
             codes.lending.to_bytes().hash256(),
             codes.oracle.to_bytes().hash256(),
