+++ dango/genesis/src/lib.rs
@@ -15,7 +15,6 @@ use {
         ResultExt, StdResult, GENESIS_SENDER,
     },
     hyperlane_types::{
-        hooks,
         isms::{self, multisig::ValidatorSet},
         mailbox::{self, Domain},
         va, Addr32,
@@ -146,11 +145,6 @@ pub fn build_rust_codes() -> Codes<ContractWrapper> {
         .with_query(Box::new(dango_dex::query))
         .build();
 
-    let fee = ContractBuilder::new(Box::new(hyperlane_fee::instantiate))
-        .with_execute(Box::new(hyperlane_fee::execute))
-        .with_query(Box::new(hyperlane_fee::query))
-        .build();
-
     let ism = ContractBuilder::new(Box::new(hyperlane_ism::instantiate))
         .with_execute(Box::new(hyperlane_ism::execute))
         .with_query(Box::new(hyperlane_ism::query))
@@ -159,11 +153,6 @@ pub fn build_rust_codes() -> Codes<ContractWrapper> {
     let mailbox = ContractBuilder::new(Box::new(hyperlane_mailbox::instantiate))
         .with_execute(Box::new(hyperlane_mailbox::execute))
         .with_query(Box::new(hyperlane_mailbox::query))
-        .build();
-
-    let merkle = ContractBuilder::new(Box::new(hyperlane_merkle::instantiate))
-        .with_execute(Box::new(hyperlane_merkle::execute))
-        .with_query(Box::new(hyperlane_merkle::query))
         .build();
 
     let va = ContractBuilder::new(Box::new(hyperlane_va::instantiate))
@@ -207,13 +196,7 @@ pub fn build_rust_codes() -> Codes<ContractWrapper> {
         account_spot,
         bank,
         dex,
-        hyperlane: Hyperlane {
-            fee,
-            ism,
-            mailbox,
-            merkle,
-            va,
-        },
+        hyperlane: Hyperlane { ism, mailbox, va },
         lending,
         oracle,
         taxman,
@@ -233,10 +216,8 @@ pub fn read_wasm_files(artifacts_dir: &Path) -> io::Result<Codes<Vec<u8>>> {
     let account_spot = fs::read(artifacts_dir.join("dango_account_spot.wasm"))?;
     let bank = fs::read(artifacts_dir.join("dango_bank.wasm"))?;
     let dex = fs::read(artifacts_dir.join("dango_dex.wasm"))?;
-    let fee = fs::read(artifacts_dir.join("hyperlane_fee.wasm"))?;
     let ism = fs::read(artifacts_dir.join("hyperlane_ism.wasm"))?;
     let mailbox = fs::read(artifacts_dir.join("hyperlane_mailbox.wasm"))?;
-    let merkle = fs::read(artifacts_dir.join("hyperlane_merkle.wasm"))?;
     let va = fs::read(artifacts_dir.join("hyperlane_va.wasm"))?;
     let lending = fs::read(artifacts_dir.join("dango_lending.wasm"))?;
     let oracle = fs::read(artifacts_dir.join("dango_oracle.wasm"))?;
@@ -251,13 +232,7 @@ pub fn read_wasm_files(artifacts_dir: &Path) -> io::Result<Codes<Vec<u8>>> {
         account_spot,
         bank,
         dex,
-        hyperlane: Hyperlane {
-            fee,
-            ism,
-            mailbox,
-            merkle,
-            va,
-        },
+        hyperlane: Hyperlane { ism, mailbox, va },
         lending,
         oracle,
         taxman,
@@ -300,10 +275,8 @@ where
     let account_spot_code_hash = upload(&mut msgs, codes.account_spot);
     let bank_code_hash = upload(&mut msgs, codes.bank);
     let dex_code_hash = upload(&mut msgs, codes.dex);
-    let hyperlane_fee_code_hash = upload(&mut msgs, codes.hyperlane.fee);
     let hyperlane_ism_code_hash = upload(&mut msgs, codes.hyperlane.ism);
     let hyperlane_mailbox_code_hash = upload(&mut msgs, codes.hyperlane.mailbox);
-    let hyperlane_merkle_code_hash = upload(&mut msgs, codes.hyperlane.merkle);
     let hyperlane_va_code_hash = upload(&mut msgs, codes.hyperlane.va);
     let lending_code_hash = upload(&mut msgs, codes.lending);
     let oracle_code_hash = upload(&mut msgs, codes.oracle);
@@ -357,24 +330,6 @@ where
         b"hyperlane/mailbox",
     );
 
-    // Instantiate Hyperlane fee hook.
-    let fee = instantiate(
-        &mut msgs,
-        hyperlane_fee_code_hash,
-        &hooks::fee::InstantiateMsg { mailbox },
-        "hyperlane/hook/fee",
-        "hyperlane/hook/fee",
-    )?;
-
-    // Instantiate Hyperlane merkle hook.
-    let merkle = instantiate(
-        &mut msgs,
-        hyperlane_merkle_code_hash,
-        &hooks::merkle::InstantiateMsg { mailbox },
-        "hyperlane/hook/merkle",
-        "hyperlane/hook/merkle",
-    )?;
-
     // Instantiate Hyperlane message ID multisig ISM.
     let ism = instantiate(
         &mut msgs,
@@ -403,8 +358,6 @@ where
             config: mailbox::Config {
                 local_domain: hyperlane_local_domain,
                 default_ism: ism,
-                default_hook: fee,
-                required_hook: merkle,
             },
         },
         "hyperlane/mailbox",
@@ -505,13 +458,7 @@ where
         account_factory,
         bank,
         dex,
-        hyperlane: Hyperlane {
-            fee,
-            ism,
-            mailbox,
-            merkle,
-            va,
-        },
+        hyperlane: Hyperlane { ism, mailbox, va },
         lending,
         oracle,
         taxman,
@@ -541,13 +488,7 @@ where
         addresses: AppAddresses {
             account_factory,
             dex,
-            hyperlane: Hyperlane {
-                fee,
-                ism,
-                mailbox,
-                merkle,
-                va,
-            },
+            hyperlane: Hyperlane { ism, mailbox, va },
             lending,
             oracle,
             warp,
