--- hyperlane/hooks/fee/src/state.rs
@@ -1,3 +1,0 @@
-use grug::{Addr, Item};
-
-pub const MAILBOX: Item<Addr> = Item::new("mailbox");
