--- hyperlane/hooks/merkle/Cargo.toml
@@ -1,27 +1,0 @@
-[package]
-authors       = { workspace = true }
-categories    = { workspace = true }
-documentation = { workspace = true }
-edition       = { workspace = true }
-license       = { workspace = true }
-name          = "hyperlane-merkle"
-repository    = { workspace = true }
-rust-version  = { workspace = true }
-version       = { workspace = true }
-
-[lib]
-crate-type = ["cdylib", "rlib"]
-
-[features]
-# If enabled, Wasm exports won't be created. This allows this contract to be
-# imported into other contracts as a library.
-library = []
-
-[dependencies]
-anyhow          = { workspace = true }
-grug            = { workspace = true }
-hyperlane-types = { workspace = true }
-
-[dev-dependencies]
-hex-literal = { workspace = true }
-sha3        = { workspace = true }
