--- hyperlane/hooks/merkle/src/lib.rs
@@ -1,5 +1,0 @@
-mod execute;
-mod query;
-mod state;
-
-pub use {execute::*, query::*, state::*};
