+++ hyperlane/mailbox/src/query.rs
@@ -1,7 +1,10 @@
 use {
-    crate::{CONFIG, DELIVERIES, NONCE},
+    crate::{CONFIG, DELIVERIES, MERKLE_TREE, NONCE},
     grug::{Hash256, ImmutableCtx, Json, JsonSerExt, StdResult},
-    hyperlane_types::mailbox::{Config, QueryMsg},
+    hyperlane_types::{
+        mailbox::{Config, QueryMsg},
+        IncrementalMerkleTree,
+    },
 };
 
 #[cfg_attr(not(feature = "library"), grug::export)]
@@ -13,6 +16,10 @@ pub fn query(ctx: ImmutableCtx, msg: QueryMsg) -> StdResult<Json> {
         },
         QueryMsg::Nonce {} => {
             let res = query_nonce(ctx)?;
+            res.to_json_value()
+        },
+        QueryMsg::Tree {} => {
+            let res = query_tree(ctx)?;
             res.to_json_value()
         },
         QueryMsg::Delivered { message_id } => {
@@ -30,6 +37,11 @@ fn query_config(ctx: ImmutableCtx) -> StdResult<Config> {
 #[inline]
 fn query_nonce(ctx: ImmutableCtx) -> StdResult<u32> {
     NONCE.current(ctx.storage)
+}
+
+#[inline]
+fn query_tree(ctx: ImmutableCtx) -> StdResult<IncrementalMerkleTree> {
+    MERKLE_TREE.load(ctx.storage)
 }
 
 #[inline]
